Apply standard naming to all Python tests
