+++ libs/pyext/test/py_u_TestCron.py
@@ -8,88 +8,105 @@
 # nor does it submit to any jurisdiction.
 #
 
+from ecflow import Alias, AttrType, Autocancel, CheckPt, ChildCmdType, Client, Clock, Cron, DState, Date, Day, Days, \
+    Defs, Ecf, Event, Expression, Family, FamilyVec, File, Flag, FlagType, FlagTypeVec, InLimit, \
+    JobCreationCtrl, Label, Late, Limit, Meter, Node, NodeContainer, NodeVec, PartExpression, PrintStyle, \
+    Repeat, RepeatDate, RepeatDay, RepeatEnumerated, RepeatInteger, RepeatString, SState, State, Style, \
+    Submittable, Suite, SuiteVec, Task, TaskVec, Time, TimeSeries, TimeSlot, Today, UrlCmd, Variable, \
+    VariableList, Verify, WhyCmd, ZombieAttr, ZombieType, ZombieUserActionType, Trigger, Complete, Edit, Defstatus
+import unittest
+import sys
 import os
-import ecflow as ecf
-import itertools as it
-
-import ecflow_test_util as Test
 
 
-def can_create_cron_from_default_parameters():
-    cron = ecf.Cron()
+class TestCron(unittest.TestCase):
+    def setUp(self):
+        cron = Cron()
+        cron.set_week_days([0, 1, 2, 3, 4, 5, 6])
+        cron.set_days_of_month([1, 2, 3, 4, 5, 6])
+        cron.set_months([1, 2, 3, 4, 5, 6])
+        cron.set_time_series("+00:00 23:00 00:30")
+        self.cron1 = cron
 
-    cron.set_days_of_month([10, 20, 30])
-    cron.set_week_days([0, 1, 2, 3, 4, 5, 6])
-    cron.set_months([1, 2, 3, 4, 5, 6])
+        cron2 = Cron()
+        cron2.set_time_series(1, 30)  # default relative = false, added in release 4.0.7
+        self.cron2 = cron2
 
-    assert cron.days_of_month == [10, 20, 30]
-    assert cron.week_days == [0, 1, 2, 3, 4, 5, 6]
-    assert cron.months == [1, 2, 3, 4, 5, 6]
+        cron3 = Cron()
+        cron3.set_week_days([0, 1, 2, 3, 4, 5, 6])
+        cron3.set_time_series(1, 30, True)
+        self.cron3 = cron3
 
+        cron4 = Cron()
+        cron4.set_week_days([0, 1, 2, 3, 4, 5, 6])
+        cron4.set_time_series("00:30 01:30 00:01")
+        self.cron4 = cron4
 
-def can_create_cron_that_runs_every_day_at_2pm():
-    cron = ecf.Cron('14:00')
-    assert cron.time() == ecf.TimeSeries(14, 00)
+        cron5 = Cron()
+        cron5.set_week_days([0, 1, 2, 3, 4, 5, 6])
+        cron5.set_time_series("+00:30")
+        self.cron5 = cron5
 
+        start = TimeSlot(0, 0)
+        finish = TimeSlot(23, 0)
+        incr = TimeSlot(0, 30)
+        time_series = TimeSeries(start, finish, incr, True)
+        cron6 = Cron(time_series)
+        cron6.set_week_days([0, 1, 2, 3])
+        self.cron6 = cron6
 
-def can_create_cron_that_runs_every_30_minutes_between_0am_and_8pm_on_the_first_six_days_of_the_months_from_january_until_july():
-    cron = ecf.Cron('+00:00 20:00 00:30', days_of_month=[1, 2, 3, 4, 5, 6], months=[1, 2, 3, 4, 5, 6])
+        cron7 = Cron()
+        cron7.set_week_days([0, 1, 2])  # This cant overlap with last_week_days_of_the_month
+        cron7.set_last_week_days_of_the_month([3, 4, 5, 6])  # This cant overlap with days_of_week
+        cron7.set_days_of_month([1, 2, 3, 4, 5, 6])
+        cron7.set_months([1, 2, 3, 4, 5, 6])
+        cron7.set_time_series("+00:00 23:00 00:30")
+        cron7.set_last_day_of_the_month()
+        self.cron7 = cron7
 
-    assert cron.time() == ecf.TimeSeries(ecf.TimeSlot(0, 0), ecf.TimeSlot(20, 0), ecf.TimeSlot(0, 30), True)
-    assert cron.days_of_month == [1, 2, 3, 4, 5, 6]
-    assert cron.months == [1, 2, 3, 4, 5, 6]
+    def test_me0(self):
+        cron = Cron("+00:00 23:00 00:30", days_of_week=[0, 1, 2, 3, 4, 5, 6], days_of_month=[1, 2, 3, 4, 5, 6], months=[1, 2, 3, 4, 5, 6])
+        self.assertEqual(self.cron1, cron, "cron not equal\n" + str(self.cron1) + "\n" + str(cron))
 
+        cron = Cron("01:30")
+        self.assertEqual(self.cron2, cron, "cron not equal\n" + str(self.cron2) + "\n" + str(cron))
 
-def can_create_cron_that_runs_relative_to_suite_start_time_or_task_requeue():
-    cron = ecf.Cron('+00:15 23:00 00:30', days_of_week=[0, 1, 2], days_of_month=[4, 5, 6], months=[1, 2, 3])
+        cron = Cron("+01:30", days_of_week=[0, 1, 2, 3, 4, 5, 6])
+        self.assertEqual(self.cron3, cron, "cron not equal\n" + str(self.cron3) + "\n" + str(cron))
 
-    assert cron.time() == ecf.TimeSeries(ecf.TimeSlot(0, 15), ecf.TimeSlot(23, 0), ecf.TimeSlot(0, 30), True)
-    assert cron.week_days == [0, 1, 2]
-    assert cron.days_of_month == [4, 5, 6]
-    assert cron.months == [1, 2, 3]
+        cron = Cron("00:30 01:30 00:01", days_of_week=[0, 1, 2, 3, 4, 5, 6])
+        self.assertEqual(self.cron4, cron, "cron not equal\n" + str(self.cron4) + "\n" + str(cron))
 
-def can_create_cron_based_on_time_series():
-    bgn = ecf.TimeSlot(0 , 0)
-    end = ecf.TimeSlot(23, 0)
-    inc = ecf.TimeSlot(0, 30)
-    ts = ecf.TimeSeries(bgn, end, inc, True)
+        cron = Cron("+00:30", days_of_week=[0, 1, 2, 3, 4, 5, 6])
+        self.assertEqual(self.cron5, cron, "cron not equal\n" + str(self.cron5) + "\n" + str(cron))
 
-    cron = ecf.Cron(ts, days_of_week=[0,1,2,3,4,5,6],days_of_month=[1,2,3,4,5,6], months=[1,2])
+        cron = Cron("+00:00 23:00 00:30", days_of_week=[0, 1, 2, 3])
+        self.assertEqual(self.cron6, cron, "cron not equal\n" + str(self.cron6) + "\n" + str(cron))
 
-    assert cron.time() == ts
-    assert cron.week_days == [0,1,2,3,4,5,6]
-    assert cron.days_of_month == [1,2,3,4,5,6]
-    assert cron.months == [1,2]
+        cron = Cron("+00:00 23:00 00:30", days_of_week=[0, 1, 2],  # This cant overlap with last_week_days_of_the_month
+                    last_week_days_of_the_month=[3, 4, 5, 6],  # This cant overlap with days_of_week
+                    days_of_month=[1, 2, 3, 4, 5, 6],
+                    months=[1, 2, 3, 4, 5, 6],
+                    last_day_of_the_month=True)
+        self.assertEqual(self.cron7, cron, "cron not equal\n" + str(self.cron7) + "\n" + str(cron))
 
-def can_create_cron_with_defaults_and_set_time_series():
-    cron = ecf.Cron()
-    cron.set_time_series(1, 30, True)  # same as cron +01:30
-
-    assert cron.time() == ecf.TimeSeries(1, 30, True)
-
-def can_create_cron_that_runs_on_the_last_day_of_the_month():
-    cron = ecf.Cron()
-    cron.set_last_day_of_the_month()
-
-    assert cron.last_day_of_the_month() == True
-
-def can_create_cron_that_runs_on_the_last_weekdays_of_the_month():
-    cron = ecf.Cron()
-    cron.set_last_week_days_of_the_month([3, 4, 5])
-
-    assert cron.last_week_days_of_the_month == [3, 4, 5]
+    def test_fail(self):
+        with self.assertRaises(RuntimeError):
+            cron = Cron("");  # empty
+        with self.assertRaises(RuntimeError):
+            cron = Cron("", days_of_week=[0, 1, 2, 3, 4, 5, 6], days_of_month=[1, 2, 3, 4, 5, 6], months=[1, 2, 3, 4, 5, 6]);  # empty
+        with self.assertRaises(RuntimeError):
+            cron = Cron("0023", days_of_week=[0, 1, 2, 3, 4, 5, 6], days_of_month=[1, 2, 3, 4, 5, 6], months=[1, 2, 3, 4, 5, 6])  # bad time
+        with self.assertRaises(RuntimeError):
+            cron = Cron("00:10", days_of_weekss=[0, 1, 2, 3, 4, 5, 6], days_of_month=[1, 2, 3, 4, 5, 6], months=[1, 2, 3, 4, 5, 6])  # bad kw,days_of_weekss
+        with self.assertRaises(RuntimeError):
+            cron = Cron("00:10", days_of_week=[0, 1, 2, 3, 4, 5, 6], days_of_months=[1, 2, 3, 4, 5, 6], months=[1, 2, 3, 4, 5, 6])  # bad kw, days_of_months
+        with self.assertRaises(RuntimeError):
+            cron = Cron("00:10", days_of_week=[0, 1, 2, 3, 4, 5, 6], days_of_month=[1, 2, 3, 4, 5, 6], monthss=[1, 2, 3, 4, 5, 6])  # bad kw, monthss
+        with self.assertRaises(RuntimeError):
+            cron = Cron("00:10", days_of_week="0,1,2,3,4,5,6", days_of_month=[1, 2, 3, 4, 5, 6], month=[1, 2, 3, 4, 5, 6])  # bad kw, expect vector of string
 
 
 if __name__ == "__main__":
-    Test.print_test_start(os.path.basename(__file__))
-
-    can_create_cron_from_default_parameters()
-    can_create_cron_that_runs_every_day_at_2pm()
-    can_create_cron_that_runs_every_30_minutes_between_0am_and_8pm_on_the_first_six_days_of_the_months_from_january_until_july()
-    can_create_cron_that_runs_relative_to_suite_start_time_or_task_requeue()
-    can_create_cron_based_on_time_series()
-    can_create_cron_with_defaults_and_set_time_series()
-    can_create_cron_that_runs_on_the_last_day_of_the_month()
-    can_create_cron_that_runs_on_the_last_weekdays_of_the_month()
-
-    print("All tests pass")
+    unittest.main()
+    print("All Tests pass")
